backend/requirements.txt
# JustInsurance Student Dashboard - Backend Dependencies

# Flask and extensions
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Session==0.5.0

# HTTP client for Absorb API
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Date/time utilities
python-dateutil==2.8.2
